Linked list class made + part of stack class
Linked list class is made - not so confident in myData being a string might have to change later - a bridge we will cross later :)

Stack class is kind of made but am running into an error with types - still getting used to how they work in the context of typescript. Prevail we must.