Finished rest of Stack + all of Queue classes
Only thing that was left in the stack class was a little type bug and then i wrote the rest of the Queue. Also reorganized my files into it's own folder within Assignment 1 so that it is separated from future C++ code for this assignment. Now we test our classes and write main! The real fun stuff!!