Developing main.cpp
In the process of developing the main process for C++ first. After I am done here, we will finish the typescript portion. So far, only have written the code to read in the items into a vector, which I still need to test. We watch The Batman first then come back to it :)