C++ Compilation Set Up
Got the compilation process set up for my files, required some editing to my class definitions but it is now working. This allowed me to test the first part of the program where the lines are actually read into an array. We can now continue with reading the palindromes into the respective stacks and queues.