Linear Search Implemented
Everything is fully tested and works for linear search, complete with average comparisons. Ran into an issue where we were searching the same item 42 times over but that was because the seed value was being defined inside the loop and since the seed value was reliant on time and the loop executed fast, we had the same seed value and thus the same item 42 times. Moving it outside the loop fixed our issue. Next, binary search.