Binary Search Complete
Spent so long trying to debug my binary search algorithm just to realize I never sorted the array first so of course I wasn't going to find anything. I implemented the iterative version of binary search but kept the recursive one in case that is the one that is required for the assignment (assignment pdf doesn't specify). Hashing next!