Traversals Working
Had to make a class method that resets the processed flags to false since the first traversal flips them to true and hence needs to be flipped back to do the next traversal. Never was able to figure out the issue with the extra column of zeros with the adjacency matrix but other than that the output is correct so hopefully that is okay.